Move ArbeidsgiverNotifikasjonProdusent and friends to package producer.arbeidsgivernotifikasjon
