+++ src/main/kotlin/no/nav/syfo/BootstrapApplication.kt
@@ -20,7 +20,7 @@ import no.nav.syfo.consumer.LocalPdlConsumer
 import no.nav.syfo.consumer.syfosyketilfelle.LocalSyfosyketilfelleConsumer
 import no.nav.syfo.consumer.PdlConsumer
 import no.nav.syfo.consumer.syfosyketilfelle.SyfosyketilfelleConsumer
-import no.nav.syfo.consumer.arbeidsgiverNotifikasjonProdusent.ArbeidsgiverNotifikasjonProdusent
+import no.nav.syfo.producer.arbeidsgivernotifikasjon.ArbeidsgiverNotifikasjonProdusent
 import no.nav.syfo.consumer.dkif.DkifConsumer
 import no.nav.syfo.consumer.narmesteLeder.NarmesteLederConsumer
 import no.nav.syfo.consumer.narmesteLeder.NarmesteLederService
