+++ src/main/kotlin/no/nav/syfo/producer/arbeidsgivernotifikasjon/Constants.kt
@@ -1,4 +1,4 @@
-package no.nav.syfo.consumer.arbeidsgiverNotifikasjonProdusent
+package no.nav.syfo.producer.arbeidsgivernotifikasjon
 
 const val MUTATION_PATH_PREFIX = "/arbeidsgiver-notifikasjon-produsent-api"
 const val CREATE_NOTIFICATION_AG_MUTATION = "nyBeskjed.graphql"
