+++ src/main/kotlin/no/nav/syfo/producer/arbeidsgivernotifikasjon/OpprettNyBeskjedArbeidsgiverNotifikasjonErrorResponse.kt
@@ -1,4 +1,4 @@
-package no.nav.syfo.consumer.arbeidsgiverNotifikasjonProdusent
+package no.nav.syfo.producer.arbeidsgivernotifikasjon
 
 data class OpprettNyBeskjedArbeidsgiverNotifikasjonErrorResponse(
     val errors: List<Error>,
