+++ src/main/kotlin/no/nav/syfo/service/ArbeidsgiverNotifikasjonService.kt
@@ -1,7 +1,7 @@
 package no.nav.syfo.service
 
 import no.nav.syfo.*
-import no.nav.syfo.consumer.arbeidsgiverNotifikasjonProdusent.ArbeidsgiverNotifikasjonProdusent
+import no.nav.syfo.producer.arbeidsgivernotifikasjon.ArbeidsgiverNotifikasjonProdusent
 import no.nav.syfo.db.domain.VarselType
 import no.nav.syfo.kafka.producers.dinesykmeldte.domain.ArbeidsgiverNotifikasjon
 import java.time.LocalDateTime
